+++ CHANGELOG.md
@@ -13,6 +13,20 @@ For more details or to discuss releases, please visit the
 
 - change default login to `admin`/`admin` (used to be `admin@admin.com`, but
   there was no reason to have a bogus email address). (#730)
+- file client/node
+  - option to store binary files
+  - display filename, file size, and stored size
+  - create file client backend code that runs for file nodes
+  - calculate and populate md5sum when file contents change
+  - display md5sum in file node UI
+- serial client/node
+  - add serial file download -- can be used for MCU updates
+  - fix issues with populate node ID for high rate data
+- db client
+  - fix crash if node ID is not populated correctly in data
+- client (BREAKING API CHANGE)
+  - renamed client.Group -> client.RunGroup. This is so we don't conflict with
+    the client that manages group nodes
 
 ## [[0.17.0] - 2024-08-05](https://github.com/simpleiot/simpleiot/releases/tag/v0.17.0)